+++ Source/ACE.Entity/AceObject.cs
@@ -267,21 +267,21 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? PhysicsState
+        public int? PhysicsState
         {
             get { return GetIntProperty(PropertyInt.PhysicsState); }
             set { SetIntProperty(PropertyInt.PhysicsState, value); }
         }
 
         [JsonIgnore]
-        public uint? WeenieType
+        public int? WeenieType
         {
             get { return GetIntProperty(PropertyInt.WeenieType); }
             set { SetIntProperty(PropertyInt.WeenieType, value); }
         }
 
         [JsonIgnore]
-        public uint? ItemType
+        public int? ItemType
         {
             get { return GetIntProperty(PropertyInt.ItemType); }
             set { SetIntProperty(PropertyInt.ItemType, value); }
@@ -312,42 +312,42 @@ namespace ACE.Entity
         public byte? ItemsCapacity
         {
             get { return (byte?)GetIntProperty(PropertyInt.ItemsCapacity); }
-            set { SetIntProperty(PropertyInt.ItemsCapacity, (uint)value); }
+            set { SetIntProperty(PropertyInt.ItemsCapacity, (int)value); }
         }
 
         [JsonIgnore]
         public byte? ContainersCapacity
         {
             get { return (byte?)GetIntProperty(PropertyInt.ContainersCapacity); }
-            set { SetIntProperty(PropertyInt.ContainersCapacity, (uint)value); }
+            set { SetIntProperty(PropertyInt.ContainersCapacity, (int)value); }
         }
 
         [JsonIgnore]
-        public uint? AmmoType
+        public int? AmmoType
         {
             get { return GetIntProperty(PropertyInt.AmmoType); }
-            set { SetIntProperty(PropertyInt.AmmoType, (uint)value); }
+            set { SetIntProperty(PropertyInt.AmmoType, (int)value); }
         }
 
         [JsonIgnore]
-        public uint? Value
+        public int? Value
         {
             get { return GetIntProperty(PropertyInt.Value); }
             set { SetIntProperty(PropertyInt.Value, value); }
         }
 
         [JsonIgnore]
-        public uint? UseCreateContractId
+        public int? UseCreateContractId
         {
             get { return GetIntProperty(PropertyInt.UseCreatesContractId); }
             set { SetIntProperty(PropertyInt.UseCreatesContractId, value); }
         }
 
         [JsonIgnore]
-        public uint? ItemUseable
+        public int? ItemUseable
         {
             get { return GetIntProperty(PropertyInt.ItemUseable); }
-            set { SetIntProperty(PropertyInt.ItemUseable, (uint?)value); }
+            set { SetIntProperty(PropertyInt.ItemUseable, (int)value); }
         }
 
         [JsonIgnore]
@@ -358,14 +358,14 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? TargetType
+        public int? TargetType
         {
             get { return GetIntProperty(PropertyInt.TargetType); }
             set { SetIntProperty(PropertyInt.TargetType, value); }
         }
 
         [JsonIgnore]
-        public uint? UiEffects
+        public int? UiEffects
         {
             get { return GetIntProperty(PropertyInt.UiEffects); }
             set { SetIntProperty(PropertyInt.UiEffects, value); }
@@ -379,7 +379,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? DefaultCombatStyle
+        public int? DefaultCombatStyle
         {
             get { return GetIntProperty(PropertyInt.DefaultCombatStyle); }
             set { SetIntProperty(PropertyInt.DefaultCombatStyle, value); }
@@ -424,7 +424,7 @@ namespace ACE.Entity
         /// If in a container, then what slot am I currently in.
         /// </summary>
         [JsonIgnore]
-        public uint? Placement
+        public int? Placement
         {
             get { return GetIntProperty(PropertyInt.Placement); }
             set { SetIntProperty(PropertyInt.Placement, value); }
@@ -445,21 +445,21 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? ValidLocations
+        public int? ValidLocations
         {
             get { return GetIntProperty(PropertyInt.ValidLocations); }
             set { SetIntProperty(PropertyInt.ValidLocations, value); }
         }
 
         [JsonIgnore]
-        public uint? CurrentWieldedLocation
+        public int? CurrentWieldedLocation
         {
             get { return GetIntProperty(PropertyInt.CurrentWieldedLocation); }
             set { SetIntProperty(PropertyInt.CurrentWieldedLocation, value); }
         }
 
         [JsonIgnore]
-        public uint? ClothingPriority
+        public int? ClothingPriority
         {
             get { return GetIntProperty(PropertyInt.ClothingPriority); }
             set { SetIntProperty(PropertyInt.ClothingPriority, value); }
@@ -487,7 +487,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? ItemWorkmanship
+        public int? ItemWorkmanship
         {
             get { return GetIntProperty(PropertyInt.ItemWorkmanship); }
             set { SetIntProperty(PropertyInt.ItemWorkmanship, value); }
@@ -543,7 +543,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? SharedCooldown
+        public int? SharedCooldown
         {
             get { return GetIntProperty(PropertyInt.SharedCooldown); }
             set { SetIntProperty(PropertyInt.SharedCooldown, value); }
@@ -565,7 +565,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? ParentLocation
+        public int? ParentLocation
         {
             get { return GetIntProperty(PropertyInt.ParentLocation); }
             set { SetIntProperty(PropertyInt.ParentLocation, value); }
@@ -593,7 +593,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? PlacementPosition
+        public int? PlacementPosition
         {
             get { return GetIntProperty(PropertyInt.PlacementPosition); }
             set { SetIntProperty(PropertyInt.PlacementPosition, value); }
@@ -628,21 +628,21 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? CharacterOptions1
+        public int? CharacterOptions1
         {
             get { return GetIntProperty(PropertyInt.CharacterOptions1); }
             set { SetIntProperty(PropertyInt.CharacterOptions1, value); }
         }
 
         [JsonIgnore]
-        public uint? CharacterOptions2
+        public int? CharacterOptions2
         {
             get { return GetIntProperty(PropertyInt.CharacterOptions2); }
             set { SetIntProperty(PropertyInt.CharacterOptions2, value); }
         }
 
         [JsonIgnore]
-        public uint? TotalLogins
+        public int? TotalLogins
         {
             get { return GetIntProperty(PropertyInt.TotalLogins); }
             set { SetIntProperty(PropertyInt.TotalLogins, value); }
@@ -684,7 +684,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? Age
+        public int? Age
         {
             get { return GetIntProperty(PropertyInt.Age); }
             set { SetIntProperty(PropertyInt.Age, value); }
@@ -698,35 +698,35 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? AvailableSkillCredits
+        public int? AvailableSkillCredits
         {
             get { return GetIntProperty(PropertyInt.AvailableSkillCredits); }
             set { SetIntProperty(PropertyInt.AvailableSkillCredits, value); }
         }
 
         [JsonIgnore]
-        public uint? TotalSkillCredits
+        public int? TotalSkillCredits
         {
             get { return GetIntProperty(PropertyInt.TotalSkillCredits); }
             set { SetIntProperty(PropertyInt.TotalSkillCredits, value); }
         }
 
         [JsonIgnore]
-        public uint? NumDeaths
+        public int? NumDeaths
         {
             get { return GetIntProperty(PropertyInt.NumDeaths); }
             set { SetIntProperty(PropertyInt.NumDeaths, value); }
         }
 
         [JsonIgnore]
-        public uint? DeathLevel
+        public int? DeathLevel
         {
             get { return GetIntProperty(PropertyInt.DeathLevel); }
             set { SetIntProperty(PropertyInt.DeathLevel, value); }
         }
 
         [JsonIgnore]
-        public uint? VitaeCpPool
+        public int? VitaeCpPool
         {
             get { return GetIntProperty(PropertyInt.VitaeCpPool); }
             set { SetIntProperty(PropertyInt.VitaeCpPool, value); }
@@ -761,14 +761,14 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? Heritage
+        public int? Heritage
         {
             get { return GetIntProperty(PropertyInt.HeritageGroup); }
             set { SetIntProperty(PropertyInt.HeritageGroup, value); }
         }
 
         [JsonIgnore]
-        public uint? Gender
+        public int? Gender
         {
             get { return GetIntProperty(PropertyInt.Gender); }
             set { SetIntProperty(PropertyInt.Gender, value); }
@@ -859,7 +859,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? Level
+        public int? Level
         {
             get { return GetIntProperty(PropertyInt.Level); }
             set { SetIntProperty(PropertyInt.Level, value); }
@@ -936,35 +936,35 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? GeneratorTimeType
+        public int? GeneratorTimeType
         {
             get { return GetIntProperty(PropertyInt.GeneratorTimeType); }
             set { SetIntProperty(PropertyInt.GeneratorTimeType, value); }
         }
 
         [JsonIgnore]
-        public uint? GeneratorProbability
+        public int? GeneratorProbability
         {
             get { return GetIntProperty(PropertyInt.GeneratorProbability); }
             set { SetIntProperty(PropertyInt.GeneratorProbability, value); }
         }
 
         [JsonIgnore]
-        public uint? MaxGeneratedObjects
+        public int? MaxGeneratedObjects
         {
             get { return GetIntProperty(PropertyInt.MaxGeneratedObjects); }
             set { SetIntProperty(PropertyInt.MaxGeneratedObjects, value); }
         }
 
         [JsonIgnore]
-        public uint? GeneratorType
+        public int? GeneratorType
         {
             get { return GetIntProperty(PropertyInt.GeneratorType); }
             set { SetIntProperty(PropertyInt.GeneratorType, value); }
         }
 
         [JsonIgnore]
-        public uint? ActivationCreateClass
+        public int? ActivationCreateClass
         {
             get { return GetIntProperty(PropertyInt.ActivationCreateClass); }
             set { SetIntProperty(PropertyInt.ActivationCreateClass, value); }
@@ -1055,35 +1055,35 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? ResistLockpick
+        public int? ResistLockpick
         {
             get { return GetIntProperty(PropertyInt.ResistLockpick); }
             set { SetIntProperty(PropertyInt.ResistLockpick, value); }
         }
 
         [JsonIgnore]
-        public uint? AppraisalLockpickSuccessPercent
+        public int? AppraisalLockpickSuccessPercent
         {
             get { return GetIntProperty(PropertyInt.AppraisalLockpickSuccessPercent); }
             set { SetIntProperty(PropertyInt.AppraisalLockpickSuccessPercent, value); }
         }
 
         [JsonIgnore]
-        public uint? MinLevel
+        public int? MinLevel
         {
             get { return GetIntProperty(PropertyInt.MinLevel); }
             set { SetIntProperty(PropertyInt.MinLevel, value); }
         }
 
         [JsonIgnore]
-        public uint? MaxLevel
+        public int? MaxLevel
         {
             get { return GetIntProperty(PropertyInt.MaxLevel); }
             set { SetIntProperty(PropertyInt.MaxLevel, value); }
         }
 
         [JsonIgnore]
-        public uint? PortalBitmask
+        public int? PortalBitmask
         {
             get { return GetIntProperty(PropertyInt.PortalBitmask); }
             set { SetIntProperty(PropertyInt.PortalBitmask, value); }
@@ -1167,14 +1167,14 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? CharacterTitleId
+        public int? CharacterTitleId
         {
             get { return GetIntProperty(PropertyInt.CharacterTitleId); }
             set { SetIntProperty(PropertyInt.CharacterTitleId, value); }
         }
 
         [JsonIgnore]
-        public uint? NumCharacterTitles
+        public int? NumCharacterTitles
         {
             get { return GetIntProperty(PropertyInt.NumCharacterTitles); }
             set { SetIntProperty(PropertyInt.NumCharacterTitles, value); }
@@ -1440,14 +1440,14 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? ItemCurMana
+        public int? ItemCurMana
         {
             get { return GetIntProperty(PropertyInt.ItemCurMana); }
             set { SetIntProperty(PropertyInt.ItemCurMana, value); }
         }
 
         [JsonIgnore]
-        public uint? ItemMaxMana
+        public int? ItemMaxMana
         {
             get { return GetIntProperty(PropertyInt.ItemMaxMana); }
             set { SetIntProperty(PropertyInt.ItemMaxMana, value); }
@@ -1468,28 +1468,28 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? CreatureType
+        public int? CreatureType
         {
             get { return GetIntProperty(PropertyInt.CreatureType); }
             set { SetIntProperty(PropertyInt.CreatureType, value); }
         }
 
         [JsonIgnore]
-        public uint? MerchandiseItemTypes
+        public int? MerchandiseItemTypes
         {
             get { return GetIntProperty(PropertyInt.MerchandiseItemTypes); }
             set { SetIntProperty(PropertyInt.MerchandiseItemTypes, value); }
         }
 
         [JsonIgnore]
-        public uint? MerchandiseMinValue
+        public int? MerchandiseMinValue
         {
             get { return GetIntProperty(PropertyInt.MerchandiseMinValue); }
             set { SetIntProperty(PropertyInt.MerchandiseMinValue, value); }
         }
 
         [JsonIgnore]
-        public uint? MerchandiseMaxValue
+        public int? MerchandiseMaxValue
         {
             get { return GetIntProperty(PropertyInt.MerchandiseMaxValue); }
             set { SetIntProperty(PropertyInt.MerchandiseMaxValue, value); }
@@ -1526,14 +1526,14 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? AppraisalPages
+        public int? AppraisalPages
         {
             get { return GetIntProperty(PropertyInt.AppraisalPages); }
             set { SetIntProperty(PropertyInt.AppraisalPages, value); }
         }
 
         [JsonIgnore]
-        public uint? AppraisalMaxPages
+        public int? AppraisalMaxPages
         {
             get { return GetIntProperty(PropertyInt.AppraisalMaxPages); }
             set { SetIntProperty(PropertyInt.AppraisalMaxPages, value); }
@@ -1542,21 +1542,21 @@ namespace ACE.Entity
 
         // TODO: This might be wrong place to store the data being stored here.
         [JsonIgnore]
-        public uint? AvailableCharacter
+        public int? AvailableCharacter
         {
             get { return GetIntProperty(PropertyInt.AvailableCharacter); }
             set { SetIntProperty(PropertyInt.AvailableCharacter, value); }
         }
 
         [JsonIgnore]
-        public uint? Boost
+        public int? Boost
         {
             get { return GetIntProperty(PropertyInt.BoostValue); }
             set { SetIntProperty(PropertyInt.BoostValue, value); }
         }
 
         [JsonIgnore]
-        public uint? BoostEnum
+        public int? BoostEnum
         {
             get { return GetIntProperty(PropertyInt.BoosterEnum); }
             set { SetIntProperty(PropertyInt.BoosterEnum, value); }
@@ -1570,7 +1570,7 @@ namespace ACE.Entity
         }
 
         [JsonIgnore]
-        public uint? CoinValue
+        public int? CoinValue
         {
             get { return GetIntProperty(PropertyInt.CoinValue); }
             set { SetIntProperty(PropertyInt.CoinValue, value); }
@@ -1958,29 +1958,29 @@ namespace ACE.Entity
             return AceObjectPropertiesSkills.Values.ToList();
         }
 
-        public uint? GetIntProperty(PropertyInt property)
+        public int? GetIntProperty(PropertyInt property)
         {
             return IntProperties.FirstOrDefault(x => x.PropertyId == (uint)property)?.PropertyValue;
         }
 
-        protected List<uint> GetIntProperties(PropertyInt property)
+        protected List<int> GetIntProperties(PropertyInt property)
         {
             return IntProperties.Where(x => x.PropertyId == (uint)property).Where(x => x.PropertyValue != null).Select(x => x.PropertyValue.Value).ToList();
         }
 
-        protected void SetIntProperty(PropertyInt intPropertyId, uint? value)
+        protected void SetIntProperty(PropertyInt intPropertyId, int? value)
         {
             AceObjectPropertiesInt listItem = IntProperties.Find(x => x.PropertyId == (ushort)intPropertyId);
             if (value != null)
             {
                 if (listItem == null)
                 {
-                    listItem = new AceObjectPropertiesInt { PropertyId = (uint)intPropertyId, PropertyValue = (uint)value, AceObjectId = AceObjectId };
+                    listItem = new AceObjectPropertiesInt { PropertyId = (uint)intPropertyId, PropertyValue = (int)value, AceObjectId = AceObjectId };
                     IntProperties.Add(listItem);
                 }
                 else
                 {
-                    listItem.PropertyValue = (uint)value;
+                    listItem.PropertyValue = (int)value;
                 }
             }
             else