+++ Source/ACE.Entity/CachedWieldedObject.cs
@@ -12,7 +12,7 @@ namespace ACE.Entity
         [DbField("aceObjectId", (int)MySqlDbType.UInt32)]
         public uint AceObjectId { get; set; }
 
-        [DbField("wieldedLocation", (int)MySqlDbType.UInt32)]
-        public uint Placement { get; set; }
+        [DbField("wieldedLocation", (int)MySqlDbType.Int32)]
+        public int Placement { get; set; }
     }
 }