+++ Source/ACE/Network/GameAction/Actions/GameActionTrainSkill.cs
@@ -8,7 +8,7 @@ namespace ACE.Network.GameAction.Actions
         public static void Handle(ClientMessage message, Session session)
         {
             var skill = (Skill)message.Payload.ReadUInt32();
-            var creditsSpent = message.Payload.ReadUInt32();
+            var creditsSpent = message.Payload.ReadInt32();
             // train skills
             session.Player.TrainSkill(skill, creditsSpent);
         }